nayeon_commit_1116_0952_현재 pull 시 에러 발생_(common) MemberVO,CenterVO 생성_ 관련 화면 전부 수정 필요
