+++ src/main/java/com/dev/pilafix/center/member_trainer_manage/MemberManageController.java
@@ -1,0 +1,9 @@
+package com.dev.pilafix.center.member_trainer_manage;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class MemberManageController {
+	
+
+}
