+++ src/main/java/com/dev/pilafix/center/member_trainer_manage/MemberManageService.java
@@ -1,0 +1,5 @@
+package com.dev.pilafix.center.member_trainer_manage;
+
+public interface MemberManageService {
+
+}
