+++ src/main/java/com/dev/pilafix/center/member_trainer_manage/impl/MemberManageDAO.java
@@ -1,0 +1,5 @@
+package com.dev.pilafix.center.member_trainer_manage.impl;
+
+public class MemberManageDAO {
+
+}
