+++ src/main/java/com/dev/pilafix/center/member_trainer_manage/impl/MemberManageServiceImpl.java
@@ -1,0 +1,6 @@
+package com.dev.pilafix.center.member_trainer_manage.impl;
+
+public class MemberManageServiceImpl {
+	
+
+}
