+++ src/main/resources/mappings/member-login-mapper.xml
@@ -4,7 +4,7 @@
 
 
 <mapper namespace="MemberLoginDAO">
-	<resultMap type="memberlogin" id="memberloginResult">
+	<resultMap type="member" id="memberloginResult">
 		<id property="csMemberCode" column="CS_MEMBER_CODE"/>
 		<result property="csRoleCode" column="CS_ROLE_CODE" />
 		<result property="csEmailId" column="CS_EMAIL_ID" />
