+++ src/main/resources/mybatis-config.xml
@@ -12,10 +12,10 @@ PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
 		<typeAlias type="com.dev.pilafix.center.info.CenterInfoVO" alias="centerInfo" />
 		<typeAlias type="com.dev.pilafix.admin.community.CommunityVO" alias="community" />
 		<typeAlias type="com.dev.pilafix.member.community.MemberCommunityVO" alias="memberCommunity" />
-		<typeAlias type="com.dev.pilafix.admin.center_manage.CenterVO" alias="center" />
 		<typeAlias type="com.dev.pilafix.admin.terms.TermsVO" alias="terms" />
 		<typeAlias type="com.dev.pilafix.admin.center_manage.SendEmailHistoryVO" alias="emailHistory" />
-		<typeAlias type="com.dev.pilafix.member.signup.MemberVO" alias="member" />
+		<typeAlias type="com.dev.pilafix.common.member.MemberVO" alias="member" />
+		<typeAlias type="com.dev.pilafix.common.member.CenterVO" alias="center" />
 		<typeAlias type="com.dev.pilafix.member.question.QuestionVO" alias="question" />
 		<typeAlias type="com.dev.pilafix.center.question_reply.QuestionReplyVO" alias="questionReply" />
 		<typeAlias type="com.dev.pilafix.admin.faq.FaqVO" alias="faqInfo" />
@@ -26,7 +26,6 @@ PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
    		<typeAlias type="com.dev.pilafix.center.ticket.CenterTicketVO" alias="centerTicket" />
 		<typeAlias type="com.dev.pilafix.member.question.QuestionVO" alias="question" />
 		<typeAlias type="com.dev.pilafix.center.question_reply.QuestionReplyVO" alias="questionReply" />		
-		<typeAlias type="com.dev.pilafix.member.login.MemberLoginVO" alias="memberlogin" />
 		<typeAlias type="com.dev.pilafix.member.ConnectCt.ConnectCtVO" alias="ConnectCt" />
 		<typeAlias type="com.dev.pilafix.center.login.CenterLoginVO" alias="centerlogin" />
 		<typeAlias type="com.dev.pilafix.center.notice.NoticeHistoryVO" alias="noticeHistory" />
